+++ hata/discord/user/__init__.py
@@ -2,6 +2,7 @@ from .activity_change import *
 from .client_user_base import *
 from .flags import *
 from .guild_profile import *
+from .helpers import *
 from .preinstanced import *
 from .thread_profile import *
 from .user import *
@@ -15,6 +16,7 @@ __all__ = (
     *client_user_base.__all__,
     *flags.__all__,
     *guild_profile.__all__,
+    *helpers.__all__,
     *preinstanced.__all__,
     *thread_profile.__all__,
     *user.__all__,